requirements.txt
flask
flask-cors
google-generativeai
pandas
sqlalchemy
python-dotenv
requests
cloud-sql-python-connector[mssql]
pymssql
pyodbc